replaced the package with fully qualified github name and moved Xxhash specific functionalities from misleading 'crypto' package to 'hash'
