+++ bls/bls_test.go
@@ -8,9 +8,9 @@ import (
 	"math/big"
 	"testing"
 
+	"github.com/dusk-network/dusk-crypto/bn256"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"dusk-crypto/bn256"
 )
 
 func randomMessage() []byte {
