+++ go.mod
@@ -1,11 +1,10 @@
-module dusk-crypto
+module github.com/dusk-network/dusk-crypto
 
 require (
 	github.com/OneOfOne/xxhash v1.2.5
 	github.com/bwesterb/go-ristretto v1.1.0
 	github.com/pkg/errors v0.8.1
 	github.com/stretchr/testify v1.3.0
-	gitlab.dusk.network/dusk-core/dusk-go v0.0.0-20190409155819-76a6a900e485 // indirect
 	golang.org/x/crypto v0.0.0-20190513172903-22d7a77e9e5f
 	golang.org/x/sys v0.0.0-20190412213103-97732733099d
 )
