+++ hash/hash_test.go
@@ -1,8 +1,10 @@
 package hash
 
 import (
-	"crypto/rand"
+	"math/rand"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func randomMessage(size int) []byte {
@@ -26,5 +28,17 @@ func BenchmarkSha3(b *testing.B) {
 
 	for i := 0; i < b.N; i++ {
 		_, _ = Sha3256(testBytes)
+	}
+}
+
+func TestRandEntropy(t *testing.T) {
+
+	for i := 0; i < 100; i++ {
+		n := uint32(rand.Intn(1e3))
+		en, err := RandEntropy(n)
+
+		assert.Equal(t, nil, err)
+		assert.Equal(t, uint32(len(en)), uint32(n))
+
 	}
 }
