+++ merkletree/merkletree.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"errors"
 
-	"dusk-crypto/hash"
+	"github.com/dusk-network/dusk-crypto/hash"
 )
 
 // Payload is data that can be stored and checked in the Merkletree.
