+++ rangeproof/debug.go
@@ -5,11 +5,11 @@ import (
 
 	"github.com/pkg/errors"
 
-	"dusk-crypto/rangeproof/vector"
+	"github.com/dusk-network/dusk-crypto/rangeproof/vector"
 
 	ristretto "github.com/bwesterb/go-ristretto"
 
-	"dusk-crypto/rangeproof/pedersen"
+	"github.com/dusk-network/dusk-crypto/rangeproof/pedersen"
 )
 
 // Put all debug functions here
